chore: merge develop into main

+++ system/modules/admin/install/migrations/20240218223900-InboxRemoveTable.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Removes the inbox table from the Inbox module that is no longer used or supported.
  */
@@ -7,20 +8,17 @@ class InboxRemoveTable extends CmfiveMigration
 {
     public function up(): void
     {
-        //Drop the tables
-        $this->dropTable('inbox');
-        $this->dropTable('inbox_message');
-
         //Clear out the related user roles
-        $this->w->db->delete('user_role')->where('role', 'inbox_sender')->execute();
-        $this->w->db->delete('user_role')->where('role', 'inbox_reader')->execute();
+        if ($this->hasTable('user_role')) {
+            $this->w->db->delete('user_role')->where('role', 'inbox_sender')->execute();
+            $this->w->db->delete('user_role')->where('role', 'inbox_reader')->execute();
+        }
     }
 
-    public function down(): void
+    public function down(): void {}
+
+    public function description()
     {
-    }
-
-    public function description() {
         return 'Remove inbox table and roles';
     }
 }
